move long exec time test classes into separate file
