+++ tests/python/rest_api/utils.py
@@ -496,10 +496,6 @@ DATUMARO_FORMAT_FOR_DIMENSION = {
 }
 
 
-def parse_frame_step(frame_filter: str) -> int:
-    return int((frame_filter or "step=1").split("=")[1])
-
-
 def calc_end_frame(start_frame: int, stop_frame: int, frame_step: int) -> int:
     return stop_frame - ((stop_frame - start_frame) % frame_step) + frame_step
 
@@ -543,3 +539,11 @@ def invite_user_to_org(
             org_id=org_id,
         )
         return invitation
+
+
+def get_cloud_storage_content(username: str, cloud_storage_id: int, manifest: Optional[str] = None):
+    with make_api_client(username) as api_client:
+        kwargs = {"manifest_path": manifest} if manifest else {}
+
+        (data, _) = api_client.cloudstorages_api.retrieve_content_v2(cloud_storage_id, **kwargs)
+        return [f"{f['name']}{'/' if str(f['type']) == 'DIR' else ''}" for f in data["content"]]
