+++ tests/python/shared/tasks/enums.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class _SourceDataType(Enum):
+    images = "images"
+    video = "video"
