+++ tests/python/shared/tasks/utils.py
@@ -1,0 +1,3 @@
+
+def parse_frame_step(frame_filter: str) -> int:
+    return int((frame_filter or "step=1").split("=")[1])
